This is simple and fast
Use a stack, if the char is left parentheses, just push it into the stack, if it is the right parentheses, find the previous char in the stack, determine if these two matches. If true, then pop the previous one. 
Therefore, at the last, if the string is valid, there should be no char in the stack in the end. Else, it is not valid.